mdrtb: misc ui updates to new specimen managament functionality

+++ src/org/openmrs/module/mdrtb/specimen/MdrtbCulture.java
@@ -23,8 +23,8 @@ public interface MdrtbCulture extends MdrtbTest {
 	public Concept getResult();
 	public void setResult(Concept result);
 	
-	public Double getColonies();
-	public void setColonies(Double colonies);
+	public Integer getColonies();
+	public void setColonies(Integer colonies);
 	
 	public Concept getMethod();
 	public void setMethod(Concept method);
